Move websockets to src
